+++ src/util/chainUtil.ts
@@ -3,7 +3,7 @@ import { longFormatNumber, shortFormatNumber } from './numberUtil';
 export const ETHER = BigInt('1000000000000000000');
 
 export const etherToNumber = (value: bigint): number => {
-  return Number((value.valueOf() * 1000000n) / ETHER) / 1000000.0;
+  return Number((value * 1000000n) / ETHER) / 1000000.0;
 };
 
 export const shortFormatEther = (value: bigint): string => {
@@ -11,7 +11,7 @@ export const shortFormatEther = (value: bigint): string => {
   return `Ξ${numberString}`;
 };
 
-export const longFormatEther = (value: bigint): string => {
-  const numberString = longFormatNumber(etherToNumber(value));
+export const longFormatEther = (value: bigint, fractionDigits = 2): string => {
+  const numberString = longFormatNumber(etherToNumber(value), fractionDigits);
   return `Ξ${numberString}`;
 };