+++ CHANGELOG.md
@@ -9,9 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 - [MINOR] Added `updateQueryString` to `urlUtil`
+- [MINOR] Added `arrayWithRange` to `listUtil`
+- [MINOR] Added `base64DecodeUnicode` to `stringUtil`
+- [MINOR] Added `startOfDay` and `endOfDay` to `dateUtil`
 
 ### Changed
 - [MAJOR] Updated `chainUtil` to use BigInt instead of BigNumber
+- [MINOR] Updated `longFormatNumber` to have new parameters `fractionDigits`, `shouldRemoveTrailingDecimals` and `shouldAddCommas`
 
 ### Removed